feat: rename useSprintWords to useGame; make this hook more common
