+++ src/hooks/index.ts
@@ -1,12 +1,12 @@
 import useCurrentWord from './useCurrentWord';
 import useDifficulty from './useDifficulty';
+import useGame from './useGame';
 import useGrouping from './useGrouping';
 import useModal from './useModal';
 import usePagination from './usePagination';
 import useSignIn from './useSignIn';
 import useSignOut from './useSignOut';
 import useSignUp from './useSignUp';
-import useSprintWords from './useSprintWords';
 import useToggle from './useToggle';
 import useWords from './useWords';
 
@@ -21,5 +21,5 @@ export {
   useToggle,
   useSignUp,
   useSignIn,
-  useSprintWords,
+  useGame,
 };
