+++ src/types/games.ts
@@ -1,0 +1,9 @@
+import { WordsResult } from './words';
+
+export interface UpdateGameDataParams {
+  allWords: WordsResult;
+  notEasyWords: WordsResult;
+  nextWordIndex: number;
+}
+
+export type UpdateGameDataFn = (options: UpdateGameDataParams) => void;
